drivers:cc1100: change cc1100_get_channel to be transceiver API compliant

+++ drivers/cc110x/cc1100.c
@@ -683,7 +683,12 @@ char *cc1100_get_output_power(char *buf)
     return buf;
 }
 
-uint8_t cc1100_get_channel(void)
+/*
+ * @brief Get the radio channel of the device
+ *
+ * @return The current radio channel of the device.
+ */
+int32_t cc1100_get_channel(void)
 {
     return radio_channel;
 }
